docs: :ambulance: call amplitude when user is logged in
